feat/ 책은 jdbcTemplate을 사용했지만 jpa로 개발

+++ src/main/java/org/intaehwang/dddstudy/chapter10/EventStore.java
@@ -4,5 +4,5 @@ import java.util.List;
 
 public interface EventStore {
     void save(Object event);
-    List<EventEntry> get(long offset, long limit);
+    List<EventEntry> get(int offset, int limit);
 }
